checkpoint ... the spec for the command line as changed,
It turns out that we cannot transparently pass vm/launcher args to the
luancher when trying to support both script command line and
` #! /usr/bin/env -S` passing of vm arguments ... to do so will
complicate the implementation of the launcher and I need to get moving
on the implementation ... will add ` #! /usr/bin/env -S` handling at a
later date (issue #2)... here's the propsed (and not yet implemented
api):

USAGE: launcher scriptPath [ imageName [-H] [-D] [-V] [topaz vm args] -- ] [script args]
       launcher scriptPath [ imageName [-H] [-D] [-V] [topaz vm args] -- [script args] -- [non-option script args]

If topaz arguments are specified the <imageName> is required to be the first
argument following the <scriptPath>.

The launcher options may be interspersed with the topz vm args as of
GemStone 3.5.0.

If the script expects to use '--' to terminate it's argument parsing, then a
'--' must precede the script arguments on the command line.

OPTIONS
  -D
     Bring up topaz debugger for StashCommandError exceptions. Normally these exceptions result in
     an error message to sderr.
  -H
     Display help message and exit.
  -V
     Verbose output

EXAMPLES
  launcher -H                           # this message
  launcher ./hello.st solo_350 -h --    # topaz help message
  launcher ./hello.st solo_350 -- -h    # hello.st help message using solo_350 image
  launcher ./hello.st -- -h             # hello.st help message using default image
  launcher ./hello.st -h                # hello.st help message using image
  launcher ./hello.st solo_350 -D -- -x # run hello.st and bring up topaz debugger in event of a command error